backend/requirements.txt
fastapi
uvicorn
python-multipart
chromadb
sentence-transformers
pymupdf
beautifulsoup4
lxml
python-dotenv
requests
